Merge duplicate copies of the breathing curve
I want to add some customisability to the curve, so I really don't
want to have multiple copies. The workarounds for PROGMEM are a little
ugly, so I'd be very interested in hearing if there's a better way of
doing this.

I'm knew to embedded development, so I'm not sure whether moving a
static table behind a function call is considered a significant
additional overhead, so feedback on that would also be appreciated.