Add BREATHING_DEPTH to control where breathing "bottoms out"
I like my backlight breathing to be more shallow. Turning the LEDs all
the way off just feels a little to extreme and can feel distracting.
Personally, I like the breathing to bottom out around 40% of the
current backlight level, which corresponds to a BREATHING_DEPTH of
0.6.

I'm pretty happy that this scaling is done on the constant breathing
curve array at compile time, so it comes with no runtime space or time
cost.

I'm still not sure whether BREATHING_DEPTH is the clearest way to
expose this, though. I'm really open to suggestions here.